requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
youtube-transcript-api
pytube
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
langchain_openai
langchain-classic
langchain-groq
langchain_core
langgraph
groq
openai
pydantic
streamlit